test: Option, Options 도메인 패키지 통합 사항 테스트 코드에 반영
